Support nested key maps
The current logic would fail if given a nested structure when using
`#key_map`. It would attempt to go deeper in the structure with the
value returned by the block.

A user may only choose to either nest or return a transformed value. If
they wish to return a transformed value they will use `#transform`.